[PAVAN]ADD: isTrash and isArchive Features

+++ src/services/notes.service.ts
@@ -22,6 +22,12 @@ class NotesService {
     return await newNote.save();
   };
 
+  //get note by Note Id
+  public getNoteById = async (noteId: string): Promise<INote | null> => {
+    return await Note.findById(noteId); 
+  };
+
+
   // Get all notes for a user
   public getNotes = async (userId: string): Promise<INote[]> => {
     return await Note.find({ createdBy: userId });
@@ -35,13 +41,45 @@ class NotesService {
     return await Note.findByIdAndUpdate(noteId, updateData, { new: true });
   };
 
+  public trashNote = async (noteId: string): Promise<INote | null> => {
+    const doc:INote = await Note.findOne({_id:noteId});
+    if(!doc.isTrash){
+      return await Note.findByIdAndUpdate(noteId,{isTrash: true},{ new: true });
+    }
+    else{
+      return await Note.findByIdAndUpdate(noteId,{isTrash: false},{ new: true });
+    }
+    
+  };
+
+  public archiveNote= async (noteId:string) : Promise<INote | null> =>{
+    const doc:INote= await Note.findOne({_id:noteId});
+    if(doc.isArchive === false){
+      return await Note.findByIdAndUpdate(noteId,{isArchive: true},{new: true});
+    }
+    else{
+      return await Note.findByIdAndUpdate(noteId,{isArchive: false},{new: true})
+    }
+  }
+
+  public deleteNote = async (noteId: string) => {
+    try{
+      const doc:INote = await Note.findOne({_id:noteId, isTrash:true});
+      if(doc){
+        await Note.findByIdAndDelete(noteId);
+      }
+      else{
+        throw new Error("Nothing in Trash With Given Id")
+      }
+    }catch(error){
+      throw new Error("Cannot find by id and Delete: "+ error)
+    }   
+    
+  };
+
+
 
   
-
-  //Delete Notes By Note ID
-  public deleteNote = async (noteId: string): Promise<void> => {
-    await Note.findByIdAndDelete(noteId);
-  };
 }
 
 export default NotesService;
